separate source file into multiple modules
(also separates out tests)
