+++ src/otable/__init__.py
@@ -1,3 +1,5 @@
 """Table data types backed by objects."""
 
-from .otable import OColumn, OTable  # noqa
+from .ocolumn import OColumn  # noqa
+from .orow import ORow  # noqa
+from .otable import OTable  # noqa
